+++ Package.swift
@@ -3,7 +3,7 @@
 import PackageDescription
 import Foundation
 
-let versionStr = "11.5.1"
+let versionStr = "11.5.2"
 let versionPieces = versionStr.split(separator: "-")
 let versionCompontents = versionPieces[0].split(separator: ".")
 let versionExtra = versionPieces.count > 1 ? versionPieces[1] : ""
@@ -639,7 +639,6 @@ let headers: [String] = [
     "realm/util/shared_ptr.hpp",
     "realm/util/signal_blocker.hpp",
     "realm/util/string_buffer.hpp",
-    "realm/util/string_view.hpp",
     "realm/util/substitute.hpp",
     "realm/util/terminate.hpp",
     "realm/util/thread.hpp",