+++ src/realm/object-store/sync/sync_session.cpp
@@ -982,11 +982,23 @@ std::function<void()> SyncProgressNotifier::NotifierPackage::create_invocation(P
         transferrable = is_download ? current_progress.downloadable : current_progress.uploadable;
     }
     else if (captured_transferrable) {
+        if (is_download && (current_progress.downloadable == 0)) {
+            captured_transferrable = 0;
+        }
         transferrable = *captured_transferrable;
     }
     else {
-        if (is_download)
-            captured_transferrable = current_progress.downloadable;
+        if (is_download) {
+            // In protocol versions 25 and earlier, downloadable_bytes was the total
+            // size of the history. From protocol version 26, downloadable_bytes
+            // represent the non-downloaded bytes on the server. Since the user supplied
+            // progress handler interprets downloadable_bytes as the total size of
+            // downloadable bytes, this number must be calculated.  We could change the
+            // meaning of downloadable_bytes for the progress handler, but that would be
+            // a breaking change. Note that protocol version 25 (and earlier) is no
+            // longer supported by clients.
+            captured_transferrable = current_progress.downloaded + current_progress.downloadable;
+        }
         else {
             // If the sync client has not yet processed all of the local
             // transactions then the uploadable data is incorrect and we should